Rename workspacelist to userworkspacelist
